# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/user/info":{"get":{"operationId":"UserController_userInfo","summary":"","description":"دریافت اطلاعات کاربر","parameters":[{"name":"id","required":false,"in":"query","description":"شناسه","schema":{"type":"number"}},{"name":"role","required":true,"in":"query","description":"نقش","schema":{"type":"boolean"}},{"name":"address","required":true,"in":"query","description":"آدرس","schema":{"type":"boolean"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]}},"/api/v1/user/update":{"put":{"operationId":"UserController_updateUser","summary":"","description":"ویرایش کاربر","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["User"]}},"/api/v1/user/getUserRole":{"get":{"operationId":"UserController_getUserRole","summary":"","description":"دریافت لیست نقش ها","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User"]}},"/api/v1/user/addAddress":{"post":{"operationId":"UserController_addAddress","summary":"","description":"افزودن آدرس کاربر","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/userAddressAddDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAddress"}}}}},"tags":["User"]}},"/api/v1/user/removeAddress":{"delete":{"operationId":"UserController_removeaddress","summary":"","description":"حذف آدرس کاربر","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/removeAddressDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User"]}},"/api/v1/admin-user/list":{"get":{"operationId":"AdminUserController_getUserList","summary":"","description":"دریافت لیست کاربران","parameters":[{"name":"firstName","required":false,"in":"query","description":"نام ","schema":{"type":"string"}},{"name":"lastName","required":false,"in":"query","description":"نام خانوادگی","schema":{"type":"string"}},{"name":"nationalId","required":false,"in":"query","description":"کدملی","schema":{"type":"string"}},{"name":"birthDateFrom","required":false,"in":"query","description":"تاریخ تولد از","schema":{"format":"date-time","type":"string"}},{"name":"birthDateTo","required":false,"in":"query","description":"تاریخ تولد تا","schema":{"format":"date-time","type":"string"}},{"name":"cellphone","required":false,"in":"query","description":"شماره همراه","schema":{"type":"string"}},{"name":"email","required":false,"in":"query","description":"ایمیل","schema":{"type":"string"}},{"name":"username","required":false,"in":"query","description":"نام کاربری","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","description":"وضعیت","schema":{"type":"boolean"}},{"name":"search","required":false,"in":"query","description":"جست و جو","schema":{"type":"string"}},{"name":"pageIndex","required":false,"in":"query","description":"شماره صفحه","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"اندازه صفحه","schema":{"type":"number"}},{"name":"sort","required":false,"in":"query","description":"مرتب کردن براساس","schema":{"type":"string"}},{"name":"asc","required":false,"in":"query","description":"صعودی/نزولی","schema":{"type":"boolean"}},{"name":"random","required":false,"in":"query","description":"مرتب سازی رندوم","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["Admin-User"]}},"/api/v1/admin-user/info":{"get":{"operationId":"AdminUserController_userInfo","summary":"","description":"دریافت اطلاعات کاربر","parameters":[{"name":"id","required":false,"in":"query","description":"شناسه","schema":{"type":"number"}},{"name":"role","required":true,"in":"query","description":"نقش","schema":{"type":"boolean"}},{"name":"address","required":true,"in":"query","description":"آدرس","schema":{"type":"boolean"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Admin-User"]}},"/api/v1/admin-user/add":{"post":{"operationId":"AdminUserController_addUser","summary":"","description":"افزودن کاربر","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Admin-User"]}},"/api/v1/admin-user/update":{"put":{"operationId":"AdminUserController_updateUser","summary":"","description":"ویرایش کاربر","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Admin-User"]}},"/api/v1/admin-user/removeUser":{"delete":{"operationId":"AdminUserController_removeUser","summary":"","description":"حذف کاربر","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/deleteUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Admin-User"]}},"/api/v1/admin-user/getUserRole":{"get":{"operationId":"AdminUserController_getUserRole","summary":"","description":"دریافت لیست نقش ها","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Admin-User"]}},"/api/v1/crud/new":{"post":{"operationId":"CrudController_addService","summary":"","description":"","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Crud"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","summary":"","description":"ورود با نام کاربری و رمز عبور","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/signInDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/v1/auth/logout":{"get":{"operationId":"AuthController_logout","summary":"","description":"خروج","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/api/v1/auth/otpRegister":{"post":{"operationId":"AuthController_otpRegister","summary":"","description":"ثبت نام/ورود با رمز یک بار مصرف","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/otpRegisterDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/api/v1/auth/otpLogin":{"post":{"operationId":"AuthController_otpLogin","summary":"","description":"ورود با رمز یک بار مصرف","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/otpLoginDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/v1/auth/switchRole":{"post":{"operationId":"AuthController_switchRole","summary":"","description":"تغییر نقش","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/switchRoleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/api/v1/network/list":{"get":{"operationId":"NetworkController_getNetworkList","summary":"","description":"دریافت لیست شبکه","parameters":[{"name":"name","required":false,"in":"query","description":"نام","schema":{"type":"string"}},{"name":"level","required":false,"in":"query","description":"سطح","schema":{"type":"string"}},{"name":"pid","required":false,"in":"query","description":"شناسه مرجع","schema":{"type":"string"}},{"name":"search","required":false,"in":"query","description":"جست و جو","schema":{"type":"string"}},{"name":"pageIndex","required":false,"in":"query","description":"شماره صفحه","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"اندازه صفحه","schema":{"type":"number"}},{"name":"sort","required":false,"in":"query","description":"مرتب کردن براساس","schema":{"type":"string"}},{"name":"asc","required":false,"in":"query","description":"صعودی/نزولی","schema":{"type":"boolean"}},{"name":"random","required":false,"in":"query","description":"مرتب سازی رندوم","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["Network"]}},"/api/v1/network/info":{"get":{"operationId":"NetworkController_networkInfo","summary":"","description":"دریافت اطلاعات شبکه","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Network"}}}}},"tags":["Network"]}},"/api/v1/network/add":{"post":{"operationId":"NetworkController_addNetwork","summary":"","description":"افزودن شبکه","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createNetworkDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Network"]}},"/api/v1/network/update":{"put":{"operationId":"NetworkController_updateNetwork","summary":"","description":"ویرایش شبکه","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateNetworkDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Network"]}},"/api/v1/network/delete":{"delete":{"operationId":"NetworkController_deleteNetwork","summary":"","description":"حذف شبکه","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/removeNetworkDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Network"]}},"/api/v1/role/list":{"get":{"operationId":"RoleController_getRoleList","summary":"","description":"دریافت لیست نقش ها","parameters":[{"name":"name","required":false,"in":"query","description":"نام نقش","schema":{"type":"string"}},{"name":"description","required":false,"in":"query","description":"توضیحات","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","description":"فعال/غیرفعال","schema":{"type":"boolean"}},{"name":"search","required":false,"in":"query","description":"جست و جو","schema":{"type":"string"}},{"name":"pageIndex","required":false,"in":"query","description":"شماره صفحه","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"اندازه صفحه","schema":{"type":"number"}},{"name":"sort","required":false,"in":"query","description":"مرتب کردن براساس","schema":{"type":"string"}},{"name":"asc","required":false,"in":"query","description":"صعودی/نزولی","schema":{"type":"boolean"}},{"name":"random","required":false,"in":"query","description":"مرتب سازی رندوم","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["Role"]}},"/api/v1/role/info":{"get":{"operationId":"RoleController_roleInfo","summary":"","description":"دریافت اطلاعات نقش","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Role"}}}}},"tags":["Role"]}},"/api/v1/role/add":{"post":{"operationId":"RoleController_addRole","summary":"","description":"افزودن نقش","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createRoleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Role"]}},"/api/v1/role/update":{"put":{"operationId":"RoleController_updateRole","summary":"","description":"ویرایش نقش","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Role"]}},"/api/v1/role/delete":{"delete":{"operationId":"RoleController_deleteRole","summary":"","description":"حذف  نقش","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/deleteRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Role"]}},"/api/v1/role/addRoleFunc":{"post":{"operationId":"RoleController_addRoleFunc","summary":"","description":"افزودن دسترسی به نقش","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/addRoleFuncDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Role"]}},"/api/v1/func/list":{"get":{"operationId":"FuncController_getFuncList","summary":"","description":"دریافت لیست فانکشن ها","parameters":[{"name":"name","required":false,"in":"query","description":"نام/لینک","schema":{"type":"string"}},{"name":"description","required":false,"in":"query","description":"توضیحات","schema":{"type":"string"}},{"name":"service","required":false,"in":"query","description":"نام سرویس","schema":{"type":"string"}},{"name":"search","required":false,"in":"query","description":"جست و جو","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Func-فانکشن"]}},"/api/v1/func/info":{"get":{"operationId":"FuncController_funcInfo","summary":"","description":"دریافت اطلاعات فانکشن","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Func"}}}}},"tags":["Func-فانکشن"]}},"/api/v1/func/add":{"post":{"operationId":"FuncController_addFunc","summary":"","description":"افزودن فانکشن","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createFuncDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Func-فانکشن"]}},"/api/v1/func/update":{"put":{"operationId":"FuncController_updateFunc","summary":"","description":"ویرایش فانکشن","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateFuncDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Func-فانکشن"]}},"/api/v1/func/sync":{"get":{"operationId":"FuncController_funcsync","summary":"","description":"بروز رسانی لیست فانکشن ها","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Func-فانکشن"]}},"/api/v1/menu/list":{"get":{"operationId":"MenuController_getMenuList","summary":"","description":"دریافت لیست منوها","parameters":[{"name":"service","required":true,"in":"query","description":"سرویس","schema":{"enum":["Panel","App"],"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Func"}}}}}},"tags":["Menu"]}},"/api/v1/menu/info":{"get":{"operationId":"MenuController_menuInfo","summary":"","description":"دریافت اطلاعات منو","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Func"}}}}},"tags":["Menu"]}},"/api/v1/menu/add":{"post":{"operationId":"MenuController_addMenu","summary":"","description":"افزودن منو","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createMenuDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Menu"]}},"/api/v1/menu/update":{"put":{"operationId":"MenuController_updateMenu","summary":"","description":"ویرایش منو","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateMenuDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Menu"]}},"/api/v1/menu/delete":{"delete":{"operationId":"MenuController_deleteMenu","summary":"","description":"حذف منو","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/deletedMenuDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Menu"]}}},"info":{"title":"nest_base","description":"The nest_base API description","version":"1.0","contact":{}},"tags":[{"name":"nest_base","description":""}],"servers":[{"url":"http://localhost:8001"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"User":{"type":"object","properties":{}},"updateUserDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه"},"firstName":{"type":"string","description":"نام "},"lastName":{"type":"string","description":"نام خانوادگی"},"nationalId":{"type":"string","description":"کدملی"},"birthDate":{"format":"date-time","type":"string","description":"تاریخ تولد"},"cellphone":{"type":"string","description":"شماره همراه"},"email":{"type":"string","description":"ایمیل"},"username":{"type":"string","description":"نام کاربری"},"status":{"type":"boolean","description":"وضعیت"},"password":{"type":"string","description":"کلمه عبور"},"education":{"type":"string","description":"تحصیلات"},"isMarried":{"type":"boolean","description":"وضعیت تاهل"},"gender":{"type":"number","description":"جنسیت"}},"required":["id","gender"]},"userAddressAddDto":{"type":"object","properties":{"userId":{"type":"number","description":"شناسه کاربر"},"cityId":{"type":"number","description":"شناسه شهر"},"address":{"type":"string","description":"آدرس"},"postalCode":{"type":"string","description":"کدپستی"},"location":{"type":"string","description":"لوکیشن"}},"required":["cityId","address","postalCode","location"]},"UserAddress":{"type":"object","properties":{}},"removeAddressDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه آدرس"}},"required":["id"]},"createUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"نام "},"lastName":{"type":"string","description":"نام خانوادگی"},"nationalId":{"type":"string","description":"کدملی"},"birthDate":{"format":"date-time","type":"string","description":"تاریخ تولد"},"cellphone":{"type":"string","description":"شماره همراه"},"email":{"type":"string","description":"ایمیل"},"username":{"type":"string","description":"نام کاربری"},"password":{"type":"string","description":"رمز عبور"},"roleId":{"type":"number","description":"شناسه نقش"},"networkId":{"type":"number","description":"شناسه شبکه"}},"required":["cellphone","username","password","roleId","networkId"]},"deleteUserDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه کاربر"}},"required":["id"]},"signInDto":{"type":"object","properties":{"grant_type":{"enum":["password","refresh_token"],"type":"string","description":"نوع ورود"},"username":{"type":"string","description":"نام کاربری"},"password":{"type":"string","description":"کلمه عبور"},"roleId":{"type":"number","description":"شناسه نقش"},"refresh_token":{"type":"string","description":"رفرش توکن"}},"required":["grant_type"]},"otpRegisterDto":{"type":"object","properties":{"cellphone":{"type":"string","description":"شماره همراه"},"roleId":{"type":"string","description":"شناسه نقش"}},"required":["cellphone"]},"otpLoginDto":{"type":"object","properties":{"cellphone":{"type":"string","description":"شماره همراه"},"roleId":{"type":"number","description":"شناسه نقش"},"code":{"type":"number","description":"کد یک بار مصرف"}},"required":["cellphone","roleId","code"]},"switchRoleDto":{"type":"object","properties":{"userRoleId":{"type":"number","description":"شناسه نقش کاربر"}},"required":["userRoleId"]},"Network":{"type":"object","properties":{}},"createNetworkDto":{"type":"object","properties":{"name":{"type":"string","description":"نام"},"pid":{"type":"string","description":"شناسه مرجع"}},"required":["name","pid"]},"updateNetworkDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه"},"name":{"type":"string","description":"نام"},"pid":{"type":"string","description":"شناسه مرجع"}},"required":["id"]},"removeNetworkDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه"}},"required":["id"]},"Role":{"type":"object","properties":{}},"createRoleDto":{"type":"object","properties":{"name":{"type":"string","description":"نام نقش"},"description":{"type":"string","description":"توضیحات"},"status":{"type":"boolean","description":"فعال/غیرفعال"}},"required":["name"]},"updateRoleDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه نقش"},"name":{"type":"string","description":"نام نقش"},"description":{"type":"string","description":"توضیحات"},"status":{"type":"boolean","description":"فعال/غیرفعال"}},"required":["id"]},"deleteRoleDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه نقش"}},"required":["id"]},"addRoleFuncDto":{"type":"object","properties":{"roleId":{"type":"number","description":"شناسه نقش"},"funcIds":{"description":"شناسه دسترسی ها","type":"array","items":{"type":"number"}}},"required":["roleId","funcIds"]},"Func":{"type":"object","properties":{}},"createFuncDto":{"type":"object","properties":{}},"updateFuncDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه"}},"required":["id"]},"createMenuDto":{"type":"object","properties":{"link":{"type":"string","description":"نام (آدرس) منو"},"title":{"type":"string","description":" عنوان منو"},"category":{"type":"string","description":"دسته بندی منو"},"ord":{"type":"number","description":"ترتیب"},"pid":{"type":"number","description":"شناسه مرجع"},"icon":{"type":"string","description":"آیکون"},"service":{"enum":["Panel","App"],"type":"string","description":"سرویس"}},"required":["link","title","service"]},"updateMenuDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه"},"link":{"type":"string","description":"نام (آدرس) منو"},"title":{"type":"string","description":" عنوان منو"},"category":{"type":"string","description":"دسته بندی منو"},"ord":{"type":"number","description":"ترتیب"},"pid":{"type":"number","description":"شناسه مرجع"},"icon":{"type":"string","description":"آیکون"},"service":{"enum":["Panel","App"],"type":"string","description":"سرویس"}},"required":["id"]},"deletedMenuDto":{"type":"object","properties":{"id":{"type":"number","description":"سرویس"},"force":{"type":"boolean","description":"حذف به همراه زیرمجموعه"}},"required":["id"]}}}}